Added sample tests, adjusted imageTransformer denoise threshold

+++ imageTransformer/main.py
@@ -3,18 +3,19 @@ import cv2
 import os
 
 
-def image_transform(filename):
+def image_transform(filename, removeOriginals):
     imageFile = cv2.imread(filename)
     imageFile = cv2.cvtColor(imageFile, cv2.COLOR_BGR2GRAY)
     imageFile = cv2.normalize(imageFile, imageFile, 0, 255, cv2.NORM_MINMAX)
     # imageFile = remove_shadows(imageFile)
     # imageFile = cv2.convertScaleAbs(imageFile, 2, 1);
     # imageFile = cv2.fastNlMeansDenoising(imageFile, imageFile, 60.0, 7, 21)
-    imageFile = flat_denoise(imageFile, 220)
+    imageFile = flat_denoise(imageFile, 190)
     imageFile = cv2.resize(imageFile, dsize=(28, 28), interpolation=cv2.INTER_NEAREST)
 
     cv2.imwrite("%s.bmp" % os.path.splitext(filename)[0], imageFile)
-    os.remove(filename)
+    if removeOriginals:
+        os.remove(filename)
 
 
 def remove_shadows(image):
@@ -40,11 +41,11 @@ def sig(x, parameter):
     return 1 / (1 + np.exp((-parameter) * (x - 127)))
 
 
-def transformAll(location):
+def transformAll(location, removeOriginals=True):
     for address, dirs, files in os.walk(location):
         for file in files:
             if file.endswith(".png") or file.endswith(".jpg"):
-                image_transform(os.path.join(address, file))
+                image_transform(os.path.join(address, file), removeOriginals)
 
         for directory in dirs:
             transformAll(os.path.join(address, directory))
@@ -53,7 +54,7 @@ def transformAll(location):
 def main():
     folderName = ".\\dataset"
     transformAll(folderName)
-    transformAll("..\\uploaded-images")
+    transformAll("..\\uploaded-images", removeOriginals=False)
 
 
 if __name__ == '__main__':
